requirements.txt
SQLAlchemy>=1.3.15
pycryptodome>=3.9.8
scrypt>=0.8.16
